refactor(eval): move the cost report into its own file
Both this branch and PR #740 append functions to the end of
cmd/eval-conflicts/main.go, so whichever merged second hit a textual
conflict in code that has no semantic overlap. Moving the helper into
cost_report.go removes that collision; only the flag declarations remain
shared, and those are a three-line hunk rather than a hundred.
